requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite==0.19.0
async-timeout==4.0.2
attrs==23.1.0
charset-normalizer==3.1.0
chat-exporter==2.5.3
discord-ui==5.1.6
discord.py==2.2.3
emoji==2.4.0
frozenlist==1.3.3
grapheme==0.6.0
idna==3.4
multidict==6.0.4
mysql-connector-python==8.0.33
protobuf==3.20.3
pytz==2023.3
PyYAML==6.0
yarl==1.9.2
